complete Auth + Login AND +logout

+++ src/main/java/com/threeNerds/basketballDiary/interceptor/Auth.java
@@ -8,5 +8,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface Auth {
-    int GRADE() default 0;
+    long GRADE() default 0;
 }
